Change length getting from db change way of calculating distance and estimated time

+++ src/main/java/pl/edu/pw/spdb/dal/DatabaseService.java
@@ -5,5 +5,5 @@ import pl.edu.pw.spdb.model.Route;
 
 public interface DatabaseService {
     Integer getStartOrEnd(Point point, boolean isStartPoint);
-    Route findRoute(Integer startId, Integer endId, Integer maxSpeed, float distanceWeight);
+    Route findRoute(long startId, long endId, Integer maxSpeed, float distanceWeight);
 }
